1st Phase of refactoring
