+++ src/main/java/org/isf/agetype/manager/AgeTypeBrowserManager.java
@@ -112,13 +112,15 @@ public class AgeTypeBrowserManager {
 	protected void validateAgeTypes(List<AgeType> ageTypes) throws OHDataValidationException {
 		List<OHExceptionMessage> errors = new ArrayList<>();
 		for (int i = 1; i < ageTypes.size(); i++) {
-			if (ageTypes.get(i).getFrom() <= ageTypes.get(i - 1).getTo()) {
+			Boolean isOverlapping = ageTypes.get(i).getFrom() <= ageTypes.get(i - 1).getTo();
+			Boolean isGapLargerThanOne = ageTypes.get(i).getFrom() - ageTypes.get(i - 1).getTo() > 1;
+			if (isOverlapping) {
 				errors.add(
 						new OHExceptionMessage(MessageBundle.getMessage("angal.common.error.title"),
 								MessageBundle.getMessage("angal.agetype.overlappedrangespleasecheckthevalues.msg"),
 								OHSeverityLevel.ERROR));
 			}
-			if (ageTypes.get(i).getFrom() - ageTypes.get(i - 1).getTo() > 1) {
+			if (isGapLargerThanOne) {
 				errors.add(new OHExceptionMessage(MessageBundle.getMessage("angal.common.error.title"),
 						MessageBundle.getMessage("angal.agetype.somerangesarenotdefinedpleasecheckthevalues.msg"),
 						OHSeverityLevel.ERROR));
